update my events to have seperate pages
